add images from docs
